Improved mocking of package imports.
The Read-the-Doc documentation builder still failed because it cannot
import NumPy as well as SciPy and some of its subpackages, all of
which are used by this library. Adding the two libraries as build
requirements would solve this issue, but slows down the build process
considerably.

As it turned out, adding the autodoc_mock_imports directive was not
sufficient. That only affects the code documentation handled by the
autodoc Sphinx extension. But the build error occurs also because the
configuration file imports the main library directly, in an effort
to extract meta information from it, such as the version number.

The version number is declared in the libary's __init__ module.
It would be inconvenient to change that as the packaging tool Flit,
which is used to publish to PyPI, relies on that declaration, in one
central location, to keep the version number consistently up-to-date.

The external dependencies are therefore now also mocked independently
of autodoc, using the MagicMock class from the standard unittest
libary.