Removed speculation from source-code comment regarding notation.
I emailed Zdravko Botev about this. He reassured and convinced me that
this is correct in that it produces the expected results. It can be
demonstrated with an asymmetric sampling from a normal distribution
(he suggested 1000 data points) where the standard deviation for one
of the coordinates is 1/10 of the other.