Don't use dynamic package metadata.
Having dynamic metadata means that the build backend (Flit) has to build
the package even for an editable install just to access metadata such as
the version number. That is, it installs all dependencies into the build
environment and then imports the package. This is more error-prone than
it needs to be, especially when, as the case is now, the import already
fails because a dependency broke. This way, we can at least install the
package for local development, to then see the failure when running the
test suite.